+++ extension_packages/figpack_experimental/figpack_experimental/__init__.py
@@ -2,4 +2,4 @@
 figpack_experimental - Miscellaneous experimental visualizations
 """
 
-__version__ = "0.1.5"
+__version__ = "0.1.6"